Use this code when the following applies
content sectioning should be within div with an id of
main-container.

the nav bar item should have data attribute with the value the same
with the data attribute of it's sub item.

you want toggler and toggled to be active but inactive when clicking
outside of it or clicking the same toggler.